Refactor thread + corrhist bench 3350213
