+++ src/thread.rs
@@ -1,16 +1,12 @@
 use std::time::Instant;
 
-use crate::board::Board;
-use crate::history::{CaptureHistory, ContinuationHistory, CorrectionHistory, QuietHistory};
+use crate::correction::CorrectionHistories;
+use crate::history::Histories;
 use crate::moves::Move;
 use crate::network::NNUE;
-use crate::parameters::{corr_counter_weight, corr_follow_up_weight, corr_major_weight, corr_minor_weight, corr_non_pawn_weight, corr_pawn_weight};
-use crate::search::{LmrTable, Score, SearchStack};
+use crate::search::{Score, SearchStack};
 use crate::time::{LimitType, SearchLimits};
 use crate::tt::TranspositionTable;
-use crate::types::bitboard::Bitboard;
-use crate::types::piece::Piece;
-use crate::types::side::Side;
 
 pub struct ThreadData {
     pub id: usize,
@@ -20,15 +16,8 @@ pub struct ThreadData {
     pub nnue: NNUE,
     pub keys: Vec<u64>,
     pub root_ply: usize,
-    pub quiet_history: QuietHistory,
-    pub capture_history: CaptureHistory,
-    pub cont_history: ContinuationHistory,
-    pub pawn_corrhist: CorrectionHistory,
-    pub nonpawn_corrhist: [CorrectionHistory; 2],
-    pub countermove_corrhist: CorrectionHistory,
-    pub follow_up_move_corrhist: CorrectionHistory,
-    pub major_corrhist: CorrectionHistory,
-    pub minor_corrhist: CorrectionHistory,
+    pub history: Histories,
+    pub correction_history: CorrectionHistories,
     pub lmr: LmrTable,
     pub node_table: NodeTable,
     pub limits: SearchLimits,
@@ -49,15 +38,8 @@ impl Default for ThreadData {
             nnue: NNUE::default(),
             keys: Vec::new(),
             root_ply: 0,
-            quiet_history: QuietHistory::default(),
-            capture_history: CaptureHistory::default(),
-            cont_history: ContinuationHistory::default(),
-            pawn_corrhist: CorrectionHistory::default(),
-            nonpawn_corrhist: [CorrectionHistory::default(), CorrectionHistory::default()],
-            countermove_corrhist: CorrectionHistory::default(),
-            follow_up_move_corrhist: CorrectionHistory::default(),
-            major_corrhist: CorrectionHistory::default(),
-            minor_corrhist: CorrectionHistory::default(),
+            history: Histories::default(),
+            correction_history: CorrectionHistories::default(),
             lmr: LmrTable::default(),
             node_table: NodeTable::new(),
             limits: SearchLimits::new(None, None, None, None, None),
@@ -85,130 +67,8 @@ impl ThreadData {
         self.tt.clear();
         self.keys.clear();
         self.root_ply = 0;
-        self.quiet_history.clear();
-        self.capture_history.clear();
-        self.cont_history.clear();
-        self.pawn_corrhist.clear();
-        self.nonpawn_corrhist[Side::White].clear();
-        self.nonpawn_corrhist[Side::Black].clear();
-        self.countermove_corrhist.clear();
-        self.follow_up_move_corrhist.clear();
-        self.major_corrhist.clear();
-        self.minor_corrhist.clear();
-    }
-
-    pub fn is_repetition(&self, board: &Board) -> bool {
-        let curr_hash = board.hash;
-        let mut repetitions = 0;
-        let end = self.keys.len().saturating_sub(board.hm as usize + 1);
-        for ply in (end..self.keys.len().saturating_sub(2)).rev() {
-            let hash = self.keys[ply];
-            repetitions += u8::from(curr_hash == hash);
-
-            // Two-fold repetition of positions within the search tree
-            if repetitions == 1 && ply >= self.root_ply {
-                return true;
-            }
-
-            // Three-fold repetition including positions before search root
-            if repetitions == 2 {
-                return true;
-            }
-        }
-        false
-    }
-
-    pub fn update_correction_history(&mut self, board: &Board, depth: i32, ply: usize, static_eval: i32, best_score: i32) {
-        let us = board.stm;
-        let pawn_hash = board.pawn_hash;
-        let w_nonpawn_hash = board.non_pawn_hashes[Side::White];
-        let b_nonpawn_hash = board.non_pawn_hashes[Side::Black];
-
-        self.pawn_corrhist.update(us, pawn_hash, depth, static_eval, best_score);
-        self.nonpawn_corrhist[Side::White].update(us, w_nonpawn_hash, depth, static_eval, best_score);
-        self.nonpawn_corrhist[Side::Black].update(us, b_nonpawn_hash, depth, static_eval, best_score);
-        self.major_corrhist.update(us, board.major_hash, depth, static_eval, best_score);
-        self.minor_corrhist.update(us, board.minor_hash, depth, static_eval, best_score);
-
-        if ply >= 1 {
-            if let Some(prev_mv) = self.ss[ply - 1].mv {
-                let encoded_mv = prev_mv.encoded() as u64;
-                self.countermove_corrhist.update(board.stm, encoded_mv, depth, static_eval, best_score);
-            }
-        }
-        if ply >= 2 {
-            if let Some(prev_mv) = self.ss[ply - 2].mv {
-                let encoded_mv = prev_mv.encoded() as u64;
-                self.follow_up_move_corrhist.update(board.stm, encoded_mv, depth, static_eval, best_score);
-            }
-        }
-    }
-
-    #[rustfmt::skip]
-    pub fn correction(&self, board: &Board, ply: usize) -> i32 {
-
-        let pawn       = self.pawn_corrhist.get(board.stm, board.pawn_hash);
-        let white      = self.nonpawn_corrhist[Side::White].get(board.stm, board.non_pawn_hashes[Side::White]);
-        let black      = self.nonpawn_corrhist[Side::Black].get(board.stm, board.non_pawn_hashes[Side::Black]);
-        let major      = self.major_corrhist.get(board.stm, board.major_hash);
-        let minor      = self.minor_corrhist.get(board.stm, board.minor_hash);
-        let counter    = self.countermove_correction(board, ply);
-        let follow_up  = self.follow_up_move_correction(board, ply);
-
-        (pawn * 100 / corr_pawn_weight())
-            + (white * 100 / corr_non_pawn_weight())
-            + (black * 100 / corr_non_pawn_weight())
-            + (major * 100 / corr_major_weight())
-            + (minor * 100 / corr_minor_weight())
-            + (counter * 100 / corr_counter_weight())
-            + (follow_up * 100 / corr_follow_up_weight())
-
-    }
-
-    fn countermove_correction(&self, board: &Board, ply: usize) -> i32 {
-        if ply >= 1 {
-            if let Some(prev_mv) = self.ss[ply - 1].mv {
-                let encoded_mv = prev_mv.encoded() as u64;
-                return self.countermove_corrhist.get(board.stm, encoded_mv);
-            }
-        }
-        0
-    }
-
-    fn follow_up_move_correction(&self, board: &Board, ply: usize) -> i32 {
-        if ply >= 2 {
-            if let Some(prev_mv) = self.ss[ply - 2].mv {
-                let encoded_mv = prev_mv.encoded() as u64;
-                return self.follow_up_move_corrhist.get(board.stm, encoded_mv);
-            }
-        }
-        0
-    }
-
-    pub fn history_score(&self, board: &Board, mv: &Move, ply: usize, threats: Bitboard, pc: Piece, captured: Option<Piece>) -> i32 {
-        if let Some(captured) = captured {
-            self.capture_history_score(board, mv, pc, captured)
-        } else {
-            self.quiet_history_score(board, mv, ply, threats)
-        }
-    }
-
-    pub fn quiet_history_score(&self, board: &Board, mv: &Move, ply: usize, threats: Bitboard) -> i32 {
-        let pc = board.piece_at(mv.from()).unwrap();
-        let quiet_score = self.quiet_history.get(board.stm, *mv, threats) as i32;
-        let mut cont_score = 0;
-        for &prev_ply in &[1, 2] {
-            if ply >= prev_ply  {
-                if let (Some(prev_mv), Some(prev_pc)) = (self.ss[ply - prev_ply].mv, self.ss[ply - prev_ply].pc) {
-                    cont_score += self.cont_history.get(prev_mv, prev_pc, mv, pc) as i32;
-                }
-            }
-        }
-        quiet_score + cont_score
-    }
-
-    pub fn capture_history_score(&self, board: &Board, mv: &Move, pc: Piece, captured: Piece) -> i32 {
-        self.capture_history.get(board.stm, pc, mv.to(), captured) as i32
+        self.history.clear();
+        self.correction_history.clear();
     }
 
     pub fn time(&self) -> u128 {
@@ -296,90 +156,32 @@ impl NodeTable {
     }
 }
 
+pub struct LmrTable {
+    table: [[i32; 64]; 256],
+}
 
-#[cfg(test)]
-mod tests {
-    use crate::board::Board;
-    use crate::moves::Move;
-    use crate::thread::ThreadData;
-
-    #[test]
-    fn test_twofold_rep_after_root() {
-        let mut td = ThreadData::default();
-        let mut board = Board::new();
-        td.keys.push(board.hash);
-        assert!(!td.is_repetition(&board));
-
-        make_move(&mut td, &mut board, "g1f3");
-        assert!(!td.is_repetition(&board));
-
-        make_move(&mut td, &mut board, "g8f6");
-        assert!(!td.is_repetition(&board));
-
-        make_move(&mut td, &mut board, "f3g1");
-        assert!(!td.is_repetition(&board));
-
-        make_move(&mut td, &mut board, "f6g8");
-        assert!(td.is_repetition(&board));
+impl LmrTable {
+    pub fn reduction(&self, depth: i32, move_count: i32) -> i32 {
+        self.table[depth.min(255) as usize][move_count.min(63) as usize]
     }
+}
 
-    #[test]
-    fn test_twofold_rep_before_root() {
-        let mut td = ThreadData::default();
-        let mut board = Board::new();
-        td.root_ply = 3;
-        td.keys.push(board.hash);
-        assert!(!td.is_repetition(&board));
+impl Default for LmrTable {
+    fn default() -> Self {
+        let base = 0.92;
+        let divisor = 3.11;
 
-        make_move(&mut td, &mut board, "g1f3");
-        assert!(!td.is_repetition(&board));
+        let mut table = [[0; 64]; 256];
 
-        make_move(&mut td, &mut board, "g8f6");
-        assert!(!td.is_repetition(&board));
+        for depth in 1..256 {
+            for move_count in 1..64 {
+                let ln_depth = (depth as f32).ln();
+                let ln_move_count = (move_count as f32).ln();
+                let reduction = (base + (ln_depth * ln_move_count / divisor)) as i32;
+                table[depth as usize][move_count as usize] = reduction;
+            }
+        }
 
-        make_move(&mut td, &mut board, "f3g1");
-        assert!(!td.is_repetition(&board));
-
-        make_move(&mut td, &mut board, "f6g8");
-        assert!(!td.is_repetition(&board));
-    }
-
-    #[test]
-    fn test_threefold_rep_before_root() {
-        let mut td = ThreadData::default();
-        let mut board = Board::new();
-        td.root_ply = 7;
-        td.keys.push(board.hash);
-        assert!(!td.is_repetition(&board));
-
-        make_move(&mut td, &mut board, "g1f3");
-        assert!(!td.is_repetition(&board));
-
-        make_move(&mut td, &mut board, "g8f6");
-        assert!(!td.is_repetition(&board));
-
-        make_move(&mut td, &mut board, "f3g1");
-        assert!(!td.is_repetition(&board));
-
-        make_move(&mut td, &mut board, "f6g8");
-        assert!(!td.is_repetition(&board));
-
-        make_move(&mut td, &mut board, "g1f3");
-        assert!(!td.is_repetition(&board));
-
-        make_move(&mut td, &mut board, "g8f6");
-        assert!(!td.is_repetition(&board));
-
-        make_move(&mut td, &mut board, "f3g1");
-        assert!(!td.is_repetition(&board));
-
-        make_move(&mut td, &mut board, "f6g8");
-        assert!(td.is_repetition(&board));
-    }
-
-    fn make_move(td: &mut ThreadData, board: &mut Board, mv: &str) {
-        let mv = Move::parse_uci(mv);
-        board.make(&mv);
-        td.keys.push(board.hash);
+        Self { table }
     }
 }
